Moved back the Account interface to its original location
